table2markdown/requirements.txt
absl-py
beautifulsoup4
jinja2
